Edinburgh NPT Network - 3km Demo Data

This data set contains network data for Edinburgh within a 3km radius, sourced from NPT. It highlights specific network characteristics important for transportation planning.

Usage

NPT_demo_3km

Format

An object of class sf (inherits from data.frame).

Examples

data(NPT_demo_3km)
head(NPT_demo_3km)
#>      all_fastest_bicycle_go_dutch
#> 2892                            0
#> 2893                            0
#> 2900                            0
#> 2901                            0
#> 2902                            0
#> 2903                            0
#>                                                                                                                                        geometry
#> 2892                                                                                                     323427.7, 323428.7, 672354.9, 672373.8
#> 2893 323427.7, 323447.7, 323450.2, 323452.1, 323453.4, 323455.2, 323456.5, 672354.9, 672353.4, 672353.4, 672354.5, 672354.4, 672355.5, 672356.6
#> 2900                                                                                                     323484.9, 323486.9, 672337.2, 672346.0
#> 2901                                                                                                     323494.3, 323495.0, 672445.0, 672448.3
#> 2902                                                                                                     323515.8, 323531.0, 672283.2, 672295.2
#> 2903                                                                                                     323515.8, 323537.4, 672283.2, 672299.5