Calculate paths from a given point to centroids within a specified distance range

This function determines the network paths from a specific point to multiple centroids based on distance thresholds. It can compute the shortest, all shortest, or all simple paths depending on the path type specified.

Usage

calculate_paths_from_point_dist(
  network,
  point,
  minDistPts = 2,
  maxDistPts = 1500,
  centroids,
  path_type = "shortest"
)

Arguments

network

An sfnetwork object representing the network.

point

An sf or sfc object containing a single point feature from which paths will be calculated.

minDistPts

The minimum distance (in meters) to consider for path calculations.

maxDistPts

The maximum distance (in meters) to consider for path calculations.

centroids

An sf object containing centroids to which paths are calculated.

path_type

A character string indicating the type of path calculation: 'shortest', 'all_shortest', or 'all_simple'. - 'shortest': Calculates the shortest path considering weights. - 'all_shortest': Calculates all paths that tie for the shortest distance, considering weights. - 'all_simple': Calculates all simple paths, ignoring weights.

Value

An sf object containing the paths that meet the criteria or NULL if no paths meet the criteria.