Create coherent network PMtiles

This function generates PMtiles for coherent network GeoJSON file using the Tippecanoe tool.

Usage

create_coherent_network_PMtiles(folder_path, city_filename, cohesive_network)

Arguments

folder_path

The directory path where the files will be saved.

city_filename

The base name for the output files, in this case using city name.

cohesive_network

An sf object representing the cohesive network.

Value

The output from the Tippecanoe command as a character vector.