Get the Edinburgh road network, within radius of 6 km of the center

Get the Edinburgh road network, within radius of 6 km of the center

Usage

get_edinburgh_6km()

Examples

get_edinburgh_6km()
#> Simple feature collection with 12096 features and 4 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 319899.7 ymin: 668038.5 xmax: 331833.5 ymax: 677990.9
#> Projected CRS: OSGB36 / British National Grid
#> # A tibble: 12,096 × 5
#>    name  road_function form_of_way road_classification                      geom
#>    <chr> <chr>         <chr>       <chr>                        <LINESTRING [m]>
#>  1 NA    Restricted L… Single Car… Unknown             (321210.7 676718.5, 3211…
#>  2 Ferr… Local Road    Single Car… Unclassified        (324017 676022, 324034 6…
#>  3 West… Local Road    Single Car… Unclassified        (329693 674189, 329672.5…
#>  4 Caiy… Local Road    Single Car… Unclassified        (324258 668416, 324255.5…
#>  5 Logi… Restricted L… Single Car… Not Classified      (325559 675412, 325547.5…
#>  6 Prin… Local Road    Single Car… Unclassified        (329750.7 670697.9, 3297…
#>  7 Morn… A Road        Single Car… A Road              (324532 671199, 324530.5…
#>  8 Rose… Local Road    Single Car… Unclassified        (323197 672989, 323207 6…
#>  9 Char… Local Road    Single Car… Unclassified        (325917 673020, 325914.5…
#> 10 Home… A Road        Collapsed … A Road              (324910.8 672963.1, 3249…
#> # ℹ 12,086 more rows