Edinburgh OpenRoads Network - 3km Demo Data

This data set contains network data for Edinburgh within a 3km radius, sourced from OpenRoads. It is formatted as a spatial data frame (sf object), suitable for spatial analysis demonstrations.

Usage

os_edinburgh_demo_3km

Format

An object of class sf (inherits from data.frame).

Examples

data(os_edinburgh_demo_3km)
head(os_edinburgh_demo_3km)
#> Simple feature collection with 6 features and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 323197 ymin: 672963.1 xmax: 325917 ymax: 676026
#> CRS:           NA
#> # A tibble: 6 × 2
#>   road_function                                                             geom
#>   <chr>                                                             <LINESTRING>
#> 1 Local Road                       (324017 676022, 324034 676024, 324051 676026)
#> 2 Restricted Local Access Road (325559 675412, 325547.5 675402.5, 325536 675393)
#> 3 Local Road                     (323197 672989, 323207 672993.5, 323217 672998)
#> 4 Local Road                   (325917 673020, 325914.5 673017.5, 325912 673015)
#> 5 A Road                                  (324910.8 672963.1, 324910.7 672963.3)
#> 6 Local Road                   (323571.4 673096.8, 323582.2 673097.3, 323593 67…